% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R-IRG_utils.R
\name{dontsleep}
\alias{dontsleep}
\title{Move the mouse so the PC does not shut.}
\usage{
dontsleep()
}
\arguments{
\item{t}{Time the function takes to move the mouse.}
}
\description{
Move the mouse so the PC does not shut.
}
\author{
Ignacio Ramos-Gutierrez
}
