% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R-IRG_utils.R
\name{getpath}
\alias{getpath}
\title{Copy directory to R understanding format.}
\usage{
getpath(x = "clipboard")
}
\arguments{
\item{x}{String. File path copied from Windows' "copy file path".}
}
\value{
An R usable file path.
}
\description{
Copy directory to R understanding format.
}
\examples{
getpath(x = "C:\Users\User\Documents")
getpath()

}
\author{
Ignacio Ramos-Gutierrez
}
