% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R-IRG_utils.R
\name{joinAttributeTable}
\alias{joinAttributeTable}
\title{Merge table to attribute table in shapefile.}
\usage{
joinAttributeTable(oldshape, table, oldshape$id, table$id)
}
\arguments{
\item{x}{Shapefile where to add the table.}

\item{y}{Table to merge into the shapefile.}

\item{xcol}{column in shapefile with id names, which must match ycol values.}

\item{ycol}{column in table with id names, which must match xcol values.}
}
\value{
A shapefile including the previous attribute table and the merged table.
}
\description{
Merge table to attribute table in shapefile.
}
\examples{
newshape <- joinAttributeTable(oldshape, table, oldshape$id, table$id)

}
\author{
Ignacio Ramos-Gutierrez, copied from internet. https://stat.ethz.ch/pipermail/r-sig-geo/2014-July/021442.html
}
