% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R-IRG_utils.R
\name{progressbar}
\alias{progressbar}
\title{Progress bar printing}
\usage{
for(i in 1:15){
  plot(1,1, main=i)
  Sys.sleep(1)
  progressbar(15, i)
 }
}
\arguments{
\item{curr.iter}{Current iteration value}

\item{tot.iter}{Number of total iterations}

\item{ini.iter}{Number of first iteration (in case loop does not start in 1. Default is 1.)}

\item{units}{units ("secs", "mins" or "hours" in which "ETC" should be calculated)}
}
\description{
Progress bar printing
}
\author{
Ignacio Ramos-Gutierrez
}
