% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R-IRG_utils.R
\name{thiswd}
\alias{thiswd}
\title{Function to set working directory where the script is}
\usage{
thiswd()

setwd(thiswd())
}
\description{
Function to set working directory where the script is
}
\author{
Ignacio Ramos-Gutierrez
}
