% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R-IRG_utils.R
\name{volunteeR}
\alias{volunteeR}
\title{Function to see if you are selected as Volunteer for the IBC}
\usage{
volunteeR()
}
\arguments{
\item{ID}{Your ID used for the lookup}
}
\description{
Function to see if you are selected as Volunteer for the IBC
}
\author{
Ignacio Ramos-Gutierrez
}
