% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R-IRG_utils.R
\name{dontsleep}
\alias{dontsleep}
\title{Move the mouse so the PC does not shut.}
\usage{
dontsleep()
}
\arguments{
\item{use}{Logical. If TRUE, the function will barely move the mouse so the
user can keep using the PC.}

\item{days}{Number of days it should be running. note that days=1 (default)
means that at 00:00 of the next day, the function will stop.}

\item{sleeptime}{time the system is slept between iterations}
}
\description{
Move the mouse so the PC does not shut.
}
\author{
Ignacio Ramos-Gutierrez
}
